requirements.txt
Pillow
tkinter
csv
